backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic[email]==2.6.4
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
python-dotenv==1.0.1
slowapi==0.1.9
